requirement.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
requests>=2.31.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
twilio>=8.0.0
razorpay>=1.3.0
aiofiles>=23.1.0
python-multipart>=0.0.6
firebase-admin>=6.5.0
pytest>=7.0.0
